+++ test/integration/test_b2_command_line.py
@@ -1994,7 +1994,6 @@ def test_replication_basic(b2_api, b2_tool, bucket_name):
             'create-bucket',
             source_bucket_name,
             'allPublic',
-            '--fileLockEnabled',
             '--replication',
             source_replication_configuration_json,
             *get_bucketinfo(),
@@ -2008,6 +2007,12 @@ def test_replication_basic(b2_api, b2_tool, bucket_name):
 
     # test that source bucket is not mentioned as replication destination
     assert source_bucket['replication'].get('asReplicationDestination') is None
+
+    # ---------------- attempt enabling object lock  ----------------
+    b2_tool.should_fail(
+        ['update-bucket', source_bucket_name, '--fileLockEnabled'],
+        'ERROR: Operation not supported for buckets with source replication'
+    )
 
     # ---------------- remove replication source ----------------
 
@@ -2343,6 +2348,47 @@ def test_replication_monitoring(b2_tool, bucket_name, b2_api):
     ]
 
     b2_api.clean_bucket(source_bucket_name)
+
+
+def test_enable_file_lock_first_retention_second(b2_tool, b2_api, bucket_name):
+    # enable file lock only
+    b2_tool.should_succeed(['update-bucket', bucket_name, '--fileLockEnabled'])
+
+    # set retention with file lock already enabled
+    b2_tool.should_succeed(
+        [
+            'update-bucket', bucket_name, '--defaultRetentionMode', 'compliance',
+            '--defaultRetentionPeriod', '7 days'
+        ]
+    )
+
+    # attempt to re-enable should be a noop
+    b2_tool.should_succeed(['update-bucket', bucket_name, '--fileLockEnabled'])
+
+    b2_api.clean_bucket(bucket_name)
+
+
+def test_enable_file_lock_and_set_retention_at_once(b2_tool, b2_api, bucket_name):
+    # attempt setting retention without file lock enabled
+    b2_tool.should_fail(
+        [
+            'update-bucket', bucket_name, '--defaultRetentionMode', 'compliance',
+            '--defaultRetentionPeriod', '7 days'
+        ], 'ERROR: The bucket is not file lock enabled \(bucket_missing_file_lock\)'
+    )
+
+    # enable file lock and set retention at once
+    b2_tool.should_succeed(
+        [
+            'update-bucket', bucket_name, '--defaultRetentionMode', 'compliance',
+            '--defaultRetentionPeriod', '7 days', '--fileLockEnabled'
+        ]
+    )
+
+    # attempt to re-enable should be a noop
+    b2_tool.should_succeed(['update-bucket', bucket_name, '--fileLockEnabled'])
+
+    b2_api.clean_bucket(bucket_name)
 
 
 def _assert_file_lock_configuration(